+++ src/Columns.tsx
@@ -1,15 +1,19 @@
 import React from 'react';
 import { SelectCellFormatter } from './formatters';
 import { Column } from './types';
+import { stopPropagation } from './utils';
+
+export const SELECT_COLUMN_KEY = 'select-row';
 
 // TODO: fix type
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const SelectColumn: Column<any, any> = {
-  key: 'select-row',
+  key: SELECT_COLUMN_KEY,
   name: '',
   width: 35,
   maxWidth: 35,
   resizable: false,
+  sortable: false,
   frozen: true,
   headerRenderer(props) {
     return (
@@ -27,11 +31,22 @@ export const SelectColumn: Column<any, any> = {
         tabIndex={-1}
         isCellSelected={props.isCellSelected}
         value={props.isRowSelected}
+        onClick={stopPropagation}
         onChange={props.onRowSelectionChange}
       />
     );
   },
-  formatterOptions: {
-    focusable: true
+  groupFormatter(props) {
+    return (
+      <SelectCellFormatter
+        aria-label="Select Group"
+        tabIndex={-1}
+        isCellSelected={props.isCellSelected}
+        value={props.isRowSelected}
+        onChange={props.onRowSelectionChange}
+        // Stop propagation to prevent row selection
+        onClick={stopPropagation}
+      />
+    );
   }
 };